Revert "Merge branch 'magic_oak_uat' of https://github.com/metasfresh/metasfresh into magic_oak_uat"
(Attempt to) revert the accidental merge or master into magic_oak_uat

This reverts commit d8c4a0b7ccfa064b73cfe037e9f90b8a6832130a, reversing
changes made to dbaee5f6bf00460816dd8d5ff5fb880bb5033858.

+++ backend/de.metas.shipper.gateway.dpd/src/main/java/de/metas/shipper/gateway/dpd/DpdShipperGatewayClient.java
@@ -45,6 +45,7 @@ import de.metas.shipper.gateway.dpd.logger.DpdClientLogEvent;
 import de.metas.shipper.gateway.dpd.logger.DpdDatabaseClientLogger;
 import de.metas.shipper.gateway.dpd.model.DpdClientConfig;
 import de.metas.shipper.gateway.dpd.model.DpdOrderCustomDeliveryData;
+import de.metas.shipper.gateway.dpd.model.DpdShipperProduct;
 import de.metas.shipper.gateway.dpd.util.DpdClientUtil;
 import de.metas.shipper.gateway.dpd.util.DpdConversionUtil;
 import de.metas.shipper.gateway.dpd.util.DpdSoapHeaderWithAuth;
@@ -290,8 +291,8 @@ public class DpdShipperGatewayClient implements ShipperGatewayClient
 				//noinspection ConstantConditions
 				parcel.setContent(deliveryOrderLine.getContent());
 				parcel.setVolume(DpdConversionUtil.formatVolume(deliveryOrderLine.getPackageDimensions()));
-				parcel.setWeight(DpdConversionUtil.convertWeightKgToDag(deliveryOrderLine.getGrossWeightKg().intValue()));
-				//				parcel.setInternational(); // todo non-UE orders will be implemented in a followup task
+				parcel.setWeight(DpdConversionUtil.convertWeightKgToDag(deliveryOrderLine.getGrossWeightKg()));
+				//				parcel.setInternational(); // todo non-UE orders will be implemented in a followup task, as with dhl
 			}
 		}
 		{
